New data elements + format for current conditions.
Realized the wunderground XML service provided a lot more data than the
JSON. Started parsing from the XML and also created a transformation to
reshape the object to make more sense. Updated the readme with changes
and config.json with the new URL. Added a new dependency, so be sure to
install it.